+++ src/config/index.ts
@@ -1,6 +1,7 @@
 export { ensureGlobalPluginConfigFile } from './bootstrap';
 export * from './constants';
 export * from './council-schema';
+export * from './default-agent';
 export { deepMerge, loadAgentPrompt, loadPluginConfig } from './loader';
 export * from './schema';
 export { getAgentOverride, getCustomAgentNames } from './utils';